post create jpa

+++ src/main/java/ru/first/dryCleaning/config/PostMapper.java
@@ -11,8 +11,8 @@ public class PostMapper implements RowMapper<Post> {
     @Override
     public Post mapRow(ResultSet resultSet, int i) throws SQLException {
         Post post = new Post();
-        post.setId(resultSet.getInt("id"));
-        post.setAuthor_id(resultSet.getInt("author_id"));
+        post.setId(resultSet.getLong("id"));
+        post.setAuthor_id(resultSet.getLong("author_id"));
         post.setText(resultSet.getString("text"));
         return post;
     }
